+++ hermes/export/keras_savedmodel.py
@@ -9,8 +9,8 @@ try:
 except ImportError:
     _has_tf = False
 
-from hermes.quiver import Platform
-from hermes.quiver.exporters import Exporter
+from hermes.repo import Platform
+from hermes.export import Exporter
 
 
 class KerasSavedModelMeta(abc.ABCMeta):
