--- hermes/hermes.aeriel/hermes/aeriel/client/__init__.py
@@ -1,1 +1,0 @@
-from .client import InferenceClient
