--- hermes/hermes.aeriel/hermes/aeriel/serve/__init__.py
@@ -1,1 +1,0 @@
-from .serve import serve
