--- hermes/hermes.aeriel/pyproject.toml
@@ -1,27 +1,0 @@
-[tool.poetry]
-name = "hermes.aeriel"
-version = "0.1.0"
-description = "Triton wrapper utilities"
-authors = ["Alec Gunny <alecg@mit.edu>"]
-packages = [
-    {include = "hermes"}
-]
-# readme = "README.md"
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.12"
-tritonclient = {extras = ["all"], version = "^2.22"}
-
-numpy = {version = "^1.22", optional = true}
-spython = {version = "^0.2", optional = true}
-
-[tool.poetry.extras]
-client = ["numpy"]
-serve = ["spython"]
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^6.2"
-
-[build-system]
-requires = ["poetry>=1.2.0a2"]
-build-backend = "poetry.masonry.api"
