--- hermes/hermes.aeriel/tox.ini
@@ -1,31 +1,0 @@
-[tox]
-envlist = py{38,39,310,311}-{client,serve}
-isolated_build = true
-
-[gh-actions]
-python =
-    3.8: py38
-    3.9: py39
-    3.10: py310
-    3.11: py311
-
-[testenv]
-commands =
-    pytest {env:test_dir}
-deps = 
-    pytest>=7,<8
-
-[testenv:py{38,39,310,311}-client]
-extras = client
-setenv =
-    test_dir = tests/client
-
-[testenv:py{38,39,310,311}-serve]
-extras = serve
-setenv =
-    test_dir = tests/serve
-
-[testenv:py{38,39,310,311}-all]
-extras = [client, serve]
-setenv =
-    test_dir = tests
