--- hermes/hermes.cloudbreak/hermes/cloudbreak/clouds/base/__init__.py
@@ -1,1 +1,0 @@
-from .kubernetes import Cluster, ClusterManager, NodePool
