--- hermes/hermes.cloudbreak/hermes/cloudbreak/kubernetes/__init__.py
@@ -1,2 +1,0 @@
-from .base import DaemonSet, Deployment, Service
-from .client import K8sApiClient
