--- hermes/hermes.cloudbreak/hermes/cloudbreak/logging.py
@@ -1,4 +1,0 @@
-import logging
-
-logger = logging.getLogger("hermes.cloudbreak")
-logger.addHandler(logging.NullHandler())
