--- hermes/hermes.cloudbreak/pyproject.toml
@@ -1,32 +1,0 @@
-[tool.poetry]
-name = "hermes.cloudbreak"
-version = "0.1.0"
-description = "Cloud orchestration utilites"
-authors = ["Alec Gunny <alecg@mit.edu>"]
-packages = [
-    {include = "hermes"}
-]
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.12"
-
-# generic utilities
-rich = "^10.7"
-kubernetes = "^17.17"
-paramiko = "^2.7"
-scp = "^0.14"
-
-# google cloud utilities
-google-cloud-container = {version = "^2.7", optional = true}
-google-cloud-compute = {version = "^0.5", optional = true}
-google-cloud-storage = {version = "^1.38", optional = true}
-
-[tool.poetry.extras]
-gcp = ["google-cloud-container", "google-cloud-storage", "google-cloud-compute"]
-
-[tool.poetry.dev-dependencies]
-pytest = "^6.2"
-
-[build-system]
-requires = ["poetry>=1.2"]
-build-backend = "poetry.masonry.api"
