--- hermes/hermes.stillwater/pyproject.toml
@@ -1,26 +1,0 @@
-[tool.poetry]
-name = "hermes.stillwater"
-version = "0.1.0"
-description = "Tools for building asynchronous inference pipelines"
-authors = ["Alec Gunny <alec.gunny@gmail.com>"]
-packages = [
-    {include = "hermes"}
-]
-
-[tool.poetry.dependencies]
-python = ">=3.8,<3.12"
-tritonclient = {extras = ["all"], version = "^2.18"}
-tblib = "^1.7"
-urllib3 = "^1.26"
-
-[tool.poetry.dev-dependencies]
-pytest = "^6.2"
-pytest-depends = "^1.0"
-
-[tool.pytest.ini_options]
-log_cli = true
-log_cli_level = "debug"
-
-[build-system]
-requires = ["poetry>=1.2"]
-build-backend = "poetry.masonry.api"
