--- hermes/hermes.stillwater/tox.ini
@@ -1,16 +1,0 @@
-[tox]
-envlist = py{38,39,310,311}
-isolated_build = true
-
-[gh-actions]
-python =
-    3.8: py38
-    3.9: py39
-    3.10: py310
-    3.11: py311
-
-[testenv:py{38,39,310,311}]
-commands = 
-    pytest tests
-deps = 
-    pytest>=7,<8
