+++ pyproject.toml
@@ -1,13 +1,33 @@
 [tool.poetry]
 name = "hermes"
-version = "0.0.1"
+version = "0.1.0"
 description = "Inference-as-a-Service deployment made simple"
 authors = [
     "Alec Gunny <alec.gunny@ligo.org>",
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.11"
+python = ">=3.9,<3.12"
+tritonclient = {extras = ["all"], version = "^2.22"}
+protobuf = "^3.17"
+
+
+# client library dependencies
+[tool.poetry.group.client]
+optional = true
+
+[tool.poetry.group.client.dependencies]
+numpy = "^1.22"
+requests = "^2.26"
+
+
+# simple serving wrapper dependencies
+[tool.poetry.group.serve]
+optional = true
+
+[tool.poetry.group.serve.dependencies]
+spython = "^0.2"
+
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^2.16"
@@ -31,5 +51,5 @@ multi_line_output = 3
 include_trailing_comma = true
 
 [build-system]
-requires = ["poetry>=1.2"]
-build-backend = "poetry.masonry.api"
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
